added PatchGenerator that the world use to create patches, generatePatch in world uses patch coordinates

+++ terrainSimulator/src/TerrainPatch.h
@@ -27,7 +27,7 @@ class TerrainPatch
     GLuint* shader;
 
   public:
-    TerrainPatch(TextureData *tex, long x, long y, GLuint* phongShader, char *imagePath);
+    TerrainPatch(TextureData *tex, int x, int y, GLuint* phongShader, char *imagePath);
     vec3 calcNormal(vec3 v0, vec3 v1, vec3 v2);
     float calcHeight(float x,float z,int texWidth);
     Model* geometry;
